Add PostgreSQL and Redis to docker, update socket management
This update includes the addition of PostgreSQL and Redis to the docker-compose file for more efficient data management. Socket management on the server has also been updated using a new library to handle emitting data to players. New types and utility functions have also been added to handle these changes. The interface now includes a MasterMask for the player acting as game master.